+++ Example/SourcePointMetaApp/Layers/Utility/Constants.swift
@@ -29,6 +29,8 @@ struct SPLiteral {
     static let campaign = "Campaign: "
     static let consentUUID = "ConsentUUID not available"
     static let euConsentID = "EUConsent not available"
+    static let stageEnv = "Stage"
+    static let publicEnv = "Public"
     
     static func attributedString() -> NSMutableAttributedString {
         let attributedText = NSMutableAttributedString(string: "Cookies for ", attributes: [NSAttributedString.Key.font: UIFont.systemFont(ofSize: 14)])
@@ -52,7 +54,7 @@ struct Alert
     static let messageAlreadyShown  = NSLocalizedString("There is no message matching the scenario based on the property info and device local data. \nConsider reviewing the property info or clearing the cookies. \nIf that was intended, just ignore this message.", comment: "")
     static let messageAlreadyShownOrIncorrectCampaign  = NSLocalizedString("Message will not be shown.\nEither the scenario decided that no message should be displayed or \n the property you're creating has a public|staging campaign and you set it as staging|public", comment: "")
     static let messageForEmptyTargetingParamError = NSLocalizedString("Please enter targeting parameter key and value", comment: "")
-    static let messageForDeletingPropertyData = NSLocalizedString("Do you want to delete this property?", comment: "")
+    static let messageForDeletingPropertyData = NSLocalizedString("Are you sur you want to delete this property?", comment: "")
     
     // Alert actions
     static let alert = NSLocalizedString("Alert", comment: "")